requirements.txt
Django==2.2.1
django-pipeline==1.6.14
django-csp==3.5
django-htmlmin==0.11.0
django-otp==0.6.0
mistune==0.8.4
mistune-contrib==0.1
Pygments==2.4.1
qrcode==6.1
sorl-thumbnail==12.5.0
Pillow==6.0.0
sentry-sdk==0.8.0
gunicorn==19.9.0
psycopg2-binary==2.8.2
## The following requirements were added by pip freeze:
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
html5lib==1.0.1
idna==2.8
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
requests==2.22.0
soupsieve==1.9.1
webencodings==0.5.1